# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Doc",
        "version": "1.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register Account",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Some username"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "someEmail@gmail.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "1a3t5u7p"
                                            },
                                            "password_confirmation": {
                                                "type": "string",
                                                "example": "1a3t5u7p"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log In",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "someEmail@gmail.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "1a3t5u7p"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "d520bcb82824a2638d7d63227d12daa51fde616da52be55af248d6b683009821"
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log Out",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "authToken": []
                    }
                ]
            }
        },
        "/api/comments": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Get all comments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "comments": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "post_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "author": {
                                                                "type": "string",
                                                                "example": "Some author"
                                                            },
                                                            "text": {
                                                                "type": "string",
                                                                "example": "Some text"
                                                            },
                                                            "created_at": {
                                                                "type": "timestamp",
                                                                "example": "2023-06-17T18:31:55.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "timestamp",
                                                                "example": "2023-06-17T18:31:55.000000Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "authToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Create comment",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "post_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "author": {
                                                "type": "string",
                                                "example": "Some author"
                                            },
                                            "text": {
                                                "type": "string",
                                                "example": "Some text"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "authToken": []
                    }
                ]
            }
        },
        "/api/comments/{comment}": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Get one comment",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "comment": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "post_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "Some author"
                                                },
                                                "text": {
                                                    "type": "string",
                                                    "example": "Some text"
                                                },
                                                "created_at": {
                                                    "type": "timestamp",
                                                    "example": "2023-06-17T18:31:55.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "timestamp",
                                                    "example": "2023-06-17T18:31:55.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "authToken": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comment"
                ],
                "summary": "Delete comment",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "authToken": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Comment"
                ],
                "summary": "Update comment",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "post_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "author": {
                                                "type": "string",
                                                "example": "Some author"
                                            },
                                            "text": {
                                                "type": "string",
                                                "example": "Some text"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "authToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "authToken": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}